requirements.txt
torch>=2.2.2
torchvision>=0.17.2
numpy>=1.26.4
Pillow>=10.2.0
matplotlib>=3.7.1
datasets>=2.13.0
transformers>=4.44.2
accelerate>=1.6.0
opencv-python>=4.10.0.84
tqdm>=4.66.5
facenet-pytorch>=2.6.0
Flask>=2.2.5
python-dotenv>=1.1.0
joblib
requests
Flask
gunicorn>=21.2.0

